# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Geto API",
    "description": "Geto Database Management Tool API",
    "license": {
      "name": ""
    },
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/connections": {
      "get": {
        "tags": [
          "crate::routes::connections"
        ],
        "operationId": "list_handler",
        "responses": {
          "200": {
            "description": "List all connections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Connection"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "crate::routes::connections"
        ],
        "operationId": "create_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Create connection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Connection"
                }
              }
            }
          }
        }
      }
    },
    "/api/connections/test": {
      "post": {
        "tags": [
          "crate::routes::connections"
        ],
        "operationId": "test_unsaved_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Test unsaved connection parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/connections/{id}": {
      "get": {
        "tags": [
          "crate::routes::connections"
        ],
        "operationId": "get_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get connection by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Connection"
                }
              }
            }
          },
          "404": {
            "description": "Connection not found"
          }
        }
      },
      "delete": {
        "tags": [
          "crate::routes::connections"
        ],
        "operationId": "delete_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete connection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "404": {
            "description": "Connection not found"
          }
        }
      },
      "patch": {
        "tags": [
          "crate::routes::connections"
        ],
        "operationId": "update_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Update connection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Connection"
                }
              }
            }
          },
          "404": {
            "description": "Connection not found"
          }
        }
      }
    },
...
    "/api/connections/{id}/connection-string": {
      "get": {
        "tags": [
          "crate::routes::connections"
        ],
        "operationId": "connection_string_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "withPassword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get formatted connection URL string",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionStringResponse"
                }
              }
            }
          },
          "404": {
            "description": "Connection not found"
          }
        }
      }
    },
    "/api/connections/{id}/database": {
      "post": {
        "tags": [
          "crate::routes::connections"
        ],
        "operationId": "set_database_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetDatabaseInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Switch default database",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Connection"
                }
              }
            }
          },
          "404": {
            "description": "Connection not found"
          }
        }
      }
    },
...
    "/api/connections/{id}/test": {
      "post": {
        "tags": [
          "crate::routes::connections"
        ],
        "operationId": "test_saved_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Test saved connection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestResult"
                }
              }
            }
          },
          "404": {
            "description": "Connection not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Connection": {
        "type": "object",
        "required": [
          "id",
          "name",
          "provider",
          "host",
          "port",
          "database",
          "username",
          "hasPassword",
          "sslMode",
          "readonly",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "color": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "database": {
            "type": "string"
          },
          "hasPassword": {
            "type": "boolean"
          },
          "host": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "port": {
            "type": "integer",
            "format": "int32"
          },
          "provider": {
            "type": "string"
          },
          "readonly": {
            "type": "boolean"
          },
          "ssh": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SshConfig"
              }
            ]
          },
          "sslMode": {
            "$ref": "#/components/schemas/SslMode"
          },
          "updatedAt": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "ConnectionInput": {
        "type": "object",
        "required": [
          "name",
          "host",
          "port",
          "database",
          "username"
        ],
        "properties": {
          "color": {
            "type": [
              "string",
              "null"
            ]
          },
          "database": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": [
              "string",
              "null"
            ]
          },
          "port": {
            "type": "integer",
            "format": "int32"
          },
          "provider": {
            "type": "string"
          },
          "readonly": {
            "type": "boolean"
          },
          "ssh": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SshInput"
              }
            ]
          },
          "sslMode": {
            "$ref": "#/components/schemas/SslMode"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "ConnectionStringResponse": {
        "type": "object",
        "required": [
          "connectionString"
        ],
        "properties": {
          "connectionString": {
            "type": "string"
          }
        }
      },
...
      "DeleteResponse": {
        "type": "object",
        "required": [
          "deleted"
        ],
        "properties": {
          "deleted": {
            "type": "boolean"
          }
        }
      },
...
      "SetDatabaseInput": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "SshAuthMethod": {
        "type": "string",
        "enum": [
          "password",
          "key"
        ]
      },
      "SshConfig": {
        "type": "object",
        "required": [
          "enabled",
          "host",
          "port",
          "username",
          "authMethod",
          "hasPassword",
          "hasPrivateKey",
          "hasPassphrase"
        ],
        "properties": {
          "authMethod": {
            "$ref": "#/components/schemas/SshAuthMethod"
          },
          "enabled": {
            "type": "boolean"
          },
          "hasPassphrase": {
            "type": "boolean"
          },
          "hasPassword": {
            "type": "boolean"
          },
          "hasPrivateKey": {
            "type": "boolean"
          },
          "host": {
            "type": "string"
          },
          "port": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "SshInput": {
        "type": "object",
        "required": [
          "enabled",
          "host",
          "port",
          "username",
          "authMethod"
        ],
        "properties": {
          "authMethod": {
            "$ref": "#/components/schemas/SshAuthMethod"
          },
          "enabled": {
            "type": "boolean"
          },
          "host": {
            "type": "string"
          },
          "passphrase": {
            "type": [
              "string",
              "null"
            ]
          },
          "password": {
            "type": [
              "string",
              "null"
            ]
          },
          "port": {
            "type": "integer",
            "format": "int32"
          },
          "privateKey": {
            "type": [
              "string",
              "null"
            ]
          },
          "username": {
            "type": "string"
          }
        }
      },
      "SslMode": {
        "type": "string",
        "enum": [
          "disable",
          "allow",
          "prefer",
          "require",
          "verify-ca",
          "verify-full"
        ]
      },
...
      "TestResult": {
        "type": "object",
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "latencyMs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "version": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    }
  }
}